Fix some type inconsistency (doubleMin/Max aggreagator stores values as double, min/max for float type returns float value)

+++ processing/src/main/java/io/druid/query/aggregation/DoubleMinAggregatorFactory.java
@@ -167,7 +167,7 @@ public class DoubleMinAggregatorFactory extends AggregatorFactory
   @Override
   public String getTypeName()
   {
-    return "float";
+    return "double";
   }
 
   @Override
